Remove obsolete AbstractMetadataRepository methods
Three methods have been moved to the InMemoryMetadataRepository as the
 only repository still depending on these methods was the InMemory
 repository itself.

The remaining methods are shared between the Doctrine and InMemory
 repositories and my fingers are itching to get rid of the
 AbstractMetadataRepository altogether.

In this commit the unused imports have also been taken care of for all
 implementations of the MetadataRepositoryInterface.
