+++ src/OpenConext/EngineBlock/Metadata/MetadataRepository/AbstractMetadataRepository.php
@@ -2,11 +2,10 @@
 
 namespace OpenConext\EngineBlock\Metadata\MetadataRepository;
 
-use OpenConext\EngineBlock\Metadata\AttributeReleasePolicy;
 use OpenConext\EngineBlock\Metadata\Entity\AbstractRole;
 use OpenConext\EngineBlock\Metadata\Entity\IdentityProvider;
-use OpenConext\EngineBlock\Metadata\MetadataRepository\Filter\FilterInterface;
 use OpenConext\EngineBlock\Metadata\Entity\ServiceProvider;
+use OpenConext\EngineBlock\Metadata\MetadataRepository\Filter\FilterInterface;
 use OpenConext\EngineBlock\Metadata\MetadataRepository\Visitor\VisitorInterface;
 
 /**
@@ -57,70 +56,6 @@ abstract class AbstractMetadataRepository implements MetadataRepositoryInterface
     }
 
     /**
-     *
-     * WARNING: Very inefficient in-memory default.
-     *
-     * @return string[]
-     */
-    public function findAllIdentityProviderEntityIds()
-    {
-        $identityProviders = $this->findIdentityProviders();
-
-        $entityIds = array();
-        foreach ($identityProviders as $identityProvider) {
-            $entityIds[] = $identityProvider->entityId;
-        }
-
-        return $entityIds;
-    }
-
-    /**
-     *
-     * WARNING: Very inefficient in-memory default.
-     *
-     * @return string[]
-     */
-    public function findReservedSchacHomeOrganizations()
-    {
-        $schacHomeOrganizations = array();
-
-        $identityProviders = $this->findIdentityProviders();
-        foreach ($identityProviders as $identityProvider) {
-            if (!$identityProvider->schacHomeOrganization) {
-                continue;
-            }
-
-            $schacHomeOrganizations[] = $identityProvider->schacHomeOrganization;
-        }
-        return $schacHomeOrganizations;
-    }
-
-    /**
-     *
-     *
-     * WARNING: Very inefficient in-memory default.
-     *
-     * @param array $identityProviderEntityIds
-     * @return array|IdentityProvider[]
-     * @throws EntityNotFoundException
-     */
-    public function findIdentityProvidersByEntityId(array $identityProviderEntityIds)
-    {
-        $identityProviders = $this->findIdentityProviders();
-
-        $filteredIdentityProviders = array();
-        foreach ($identityProviderEntityIds as $identityProviderEntityId) {
-            if (!isset($identityProviders[$identityProviderEntityId])) {
-                // @todo warn
-                continue;
-            }
-
-            $filteredIdentityProviders[$identityProviderEntityId] = $identityProviders[$identityProviderEntityId];
-        }
-        return $filteredIdentityProviders;
-    }
-
-    /**
      * @param string $entityId
      * @return ServiceProvider
      * @throws EntityNotFoundException
@@ -153,7 +88,6 @@ abstract class AbstractMetadataRepository implements MetadataRepositoryInterface
     }
 
     /**
-     *
      * @param string $entityId
      * @return AbstractRole
      * @throws EntityNotFoundException
@@ -186,5 +120,28 @@ abstract class AbstractMetadataRepository implements MetadataRepositoryInterface
         }
 
         return null;
+    }
+
+    /**
+     * WARNING: Very inefficient in-memory default.
+     *
+     * @param array $identityProviderEntityIds
+     * @return array|IdentityProvider[]
+     * @throws EntityNotFoundException
+     */
+    public function findIdentityProvidersByEntityId(array $identityProviderEntityIds)
+    {
+        $identityProviders = $this->findIdentityProviders();
+
+        $filteredIdentityProviders = array();
+        foreach ($identityProviderEntityIds as $identityProviderEntityId) {
+            if (!isset($identityProviders[$identityProviderEntityId])) {
+                // @todo warn
+                continue;
+            }
+
+            $filteredIdentityProviders[$identityProviderEntityId] = $identityProviders[$identityProviderEntityId];
+        }
+        return $filteredIdentityProviders;
     }
 }
